back/requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
cassandra-driver==3.23.0
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.8
geomet==0.1.2
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
nbformat==5.0.6
nbimporter==0.3.2
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py4j==0.10.7
Pygments==2.6.1
pyrsistent==0.16.0
pyspark==2.4.5
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
waitress==1.4.3
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
